+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `PlaybackToggleButton` now also listens to `ON_SOURCE_LOADED` and `ON_SOURCE_UNLOADED` to properly update the playback state when the source changes
 - Update package dependencies
 
+### Fixed
+- Unnecessary line breaks in CEA-608 texts
+
 ## [2.10.4]
 
 ### Changed